# openapi-complete.yaml
openapi: 3.0.3
info:
  title: AI Guardians API - Unified Guard Processing
  description: |
    **Unified API specification for AI Guardians backend services.**
    
    This API provides comprehensive text analysis through a unified guard processing system.
    All guard services (TokenGuard, TrustGuard, ContextGuard, BiasGuard, HealthGuard) are
    processed through a single unified endpoint with intelligent routing and fallback mechanisms.
    
    ## Unified Architecture
    -  **Unified Processing** - Single endpoint for all guard services
    -  **Intelligent Routing** - Automatic service selection and load balancing
    -  **Fallback Mechanisms** - Automatic failover and service recovery
    -  **Real-time Monitoring** - Health checks and performance metrics
    -  **Enterprise Security** - JWT auth, rate limiting, CORS
    -  **Multi-Client Support** - Web, VS Code, Chrome, API clients
    
    ## Guard Services
    - **TokenGuard** - Token optimization and cost analysis
    - **TrustGuard** - Trust validation and reliability scoring
    - **ContextGuard** - Context analysis and drift detection
    - **BiasGuard** - Bias detection and mitigation
    - **HealthGuard** - System monitoring and security scanning
    
    ## Authentication
    All endpoints (except health) require JWT authentication via Bearer token in Authorization header.
    
    ## Rate Limiting
    - Free Tier: 100 requests/hour
    - Pro Tier: 1000 requests/hour
    - Enterprise: Unlimited
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  message:
                    type: string
                    example: "User registered successfully. Please verify your email."
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive JWT access and refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      description: Get a new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenData'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify email address
      description: Verify user email using token from verification email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerification'
      responses:
        '200':
          description: Email verified successfully
        '400':
          $ref: '#/components/responses/BadRequest'

  # ====================
  # TEXT ANALYSIS (3 endpoints)
  # ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
...
  schemas:
...
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
    UserRegister:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
        full_name:
          type: string
          minLength: 1
          maxLength: 255
          example: John Doe
    Token:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          example: bearer
          default: bearer
        expires_in:
          type: integer
          description: Token expiry in seconds
          example: 3600
    TokenData:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
...
    EmailVerification:
      type: object
      required: [token]
      properties:
        token:
          type: string

    # ====================
    # ANALYSIS SCHEMAS
    # ====================
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
            request_id:
              type: string
              format: uuid
              description: Request ID for tracking
            timestamp:
              type: string
              format: date-time
  responses:
    BadRequest:
      description: Bad request - Invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: invalid_request
              message: The request is invalid
              details:
                field: text
                issue: Text is required
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: authentication_failed
              message: Invalid or expired token
...